Refactor classify services and update UI components
Renamed Scanner to ClassificationService and moved ClassifyExpandWorker to the Services namespace. Enhanced IndexesPage with classification and QueuePage UI with color indicators.
