+++ src/harvest/Fip.Strive.Harvester.Web/appsettings.json
@@ -36,7 +36,7 @@
     },
     "Schedule":  {
       "HelloWorldIntervalSeconds": 5,
-      "QueueCleanupIntervalSeconds": 60
+      "QueueCleanupIntervalSeconds": 86400
     },
     "Index": {
       "DatabasePath": "E:\\strive\\data\\db\\index.db"
